change controller func

+++ app/Http/Controllers/BookController.php
@@ -16,6 +16,16 @@ class BookController extends Controller
     {
         try {
             $books = Book::all();
+            if(!$books) {
+                $res = [
+                    'status' => 'fail',
+                    'code' => 404,
+                    'message' => 'Book not found',
+                    'data' => $books
+                ];
+                return response()->json($res, 404);
+            }
+
             $res = [
                 'status' => 'success',
                 'code' => 200,
@@ -38,30 +48,21 @@ class BookController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
      * Store a newly created resource in storage.
      */
     public function store(StoreBookRequest $request)
     {
         $validated = $request->validated();
-
         try {
             $stored_book = Book::create($validated);
-
             $res = [
                 'status' => 'success',
-                'code' => 200,
-                'message' => 'return single book',
+                'code' => 201,
+                'message' => "new book's record is created",
                 'data' => $stored_book
             ];
-            return response()->json($res, 200);
+            return response()->json($res, 201);
+
         } catch (\Throwable $th) {
             $res = [
                 'status' => 'error',
@@ -81,22 +82,25 @@ class BookController extends Controller
         try {
             $book = Book::where('id', $id)->first();
 
-            if($book == null){
+            if(!$book){
                 $res = [
                     'status' => 'fail',
-                    'code' => 400,
-                    'message' => 'book record is empty',
+                    'code' => 404,
+                    'message' => 'Book not found',
                     'data' => $book
                 ];
-                return response()->json($res, 200);
+                return response()->json($res, 404);
             }
+
             $res = [
                 'status' => 'success',
                 'code' => 200,
                 'message' => 'return single book',
                 'data' => $book
             ];
+
             return response()->json($res, 200);
+
         } catch (\Throwable $th) {
             $res = [
                 'status' => 'error',
@@ -109,38 +113,44 @@ class BookController extends Controller
     }
 
     /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Book $book)
-    {
-        //
-    }
-
-    /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateBookRequest $request, Book $book)
+    public function update(UpdateBookRequest $request, $id)
     {
+        $book = Book::where('id', $id)->first();
+        if(!$book) {
+            $res = [
+                'status' => 'error',
+                'code' => 404,
+                'message' => 'Book not found, cannot update',
+            ];
+            return response()->json($res, 404);
+        }
+
         $validated = $request->validated();
+
         try {
             $updated = $book->update($validated);
 
             if(!$updated) {
                 $res = [
-                    'status' => 'error',
+                    'status' => 'fail',
                     'code' => 400,
                     'message' => 'failed to update book',
                 ];
                 return response()->json($res, 400);
             }
-            $updated_book = Book::find($book->id)->first()->toArray();
+
+            $updated_book = Book::where('id', $id)->first();
             $res = [
                 'status' => 'success',
                 'code' => 200,
                 'message' => 'return updated single book',
                 'data' => $updated_book
             ];
+
             return response()->json($res, 200);
+
         } catch (\Throwable $th) {
             $res = [
                 'status' => 'error',
@@ -155,10 +165,29 @@ class BookController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Book $book)
+    public function destroy($id)
     {
         try {
+            $book = Book::where('id', $id)->first();
+            if(!$book) {
+                $res = [
+                    'status' => 'fail',
+                    'code' => 404,
+                    'message' => 'Book not found',
+                    'data' => $book
+                ];
+                return response()->json($res, 404);
+            }
+
             $deleted = $book->delete();
+            if(!$deleted){
+                $res = [
+                    'status' => 'fail',
+                    'code' => 400,
+                    'message' => 'failed to delete book',
+                ];
+                return response()->json($res, 400);
+            }
             $res = [
                 'status' => 'success',
                 'code' => 200,
